+++ src/HW/LedgerTransportSwitchModal/index.ts
@@ -1,2 +1,1 @@
-export * from './Instructions'
 export * from './LedgerTransportSwitchModal'
