+++ src/Receive/Modals.tsx
@@ -6,7 +6,6 @@ import {useDispatch, useSelector} from 'react-redux'
 import {showErrorDialog} from '../../legacy/actions'
 import {setLedgerDeviceId, setLedgerDeviceObj} from '../../legacy/actions/hwWallet'
 import LedgerConnect from '../../legacy/components/Ledger/LedgerConnect'
-import LedgerTransportSwitchModal from '../../legacy/components/Ledger/LedgerTransportSwitchModal'
 import {Modal} from '../../legacy/components/UiKit'
 import {CONFIG} from '../../legacy/config/config'
 import {getCardanoByronConfig} from '../../legacy/config/networks'
@@ -17,6 +16,7 @@ import {errorMessages} from '../../legacy/i18n/global-messages'
 import LocalizableError from '../../legacy/i18n/LocalizableError'
 import {externalAddressIndexSelector, hwDeviceInfoSelector} from '../../legacy/selectors'
 import {Logger} from '../../legacy/utils/logging'
+import {LedgerTransportSwitchModal} from '../HW'
 import {useSelectedWallet} from '../SelectedWallet'
 import AddressModal from './AddressModal'
 import {AddressVerifyModal} from './AddressVerifyModal'
