+++ src/WalletInit/WalletInit/WalletInitScreen.tsx
@@ -4,7 +4,6 @@ import {defineMessages, useIntl} from 'react-intl'
 import {StyleSheet, View} from 'react-native'
 import {SafeAreaView} from 'react-native-safe-area-context'
 
-import LedgerTransportSwitchModal from '../../../legacy/components/Ledger/LedgerTransportSwitchModal'
 import {Button, Modal, ScreenBackground, StatusBar} from '../../../legacy/components/UiKit'
 import {CONFIG, isByron, isHaskellShelley} from '../../../legacy/config/config'
 import {isJormungandr} from '../../../legacy/config/networks'
@@ -12,6 +11,7 @@ import type {NetworkId, WalletImplementationId, YoroiProvider} from '../../../le
 import globalMessages from '../../../legacy/i18n/global-messages'
 import {WALLET_INIT_ROUTES} from '../../../legacy/RoutesList'
 import {COLORS} from '../../../legacy/styles/config'
+import {LedgerTransportSwitchModal} from '../../HW'
 import {WalletDescription} from '../WalletDescription'
 import {ExpandableItem} from './ExpandableItem'
